+++ ext/jpeg/gstjpegenc.h
@@ -58,7 +58,7 @@ struct _GstJpegEnc {
   gint format;
   gint width;
   gint height;
-  gfloat fps;
+  gdouble fps;
   /* the video buffer */
   gint bufsize;
   GstBuffer *buffer;
